druppie/mcp-servers/azure-datalake/requirements.txt
fastmcp==0.1.0
azure-storage-file-datalake==12.14.0
pyarrow==14.0.1
pandas==2.1.3
structlog==24.1.0
python-dotenv==1.0.0
